Refactor interpreter.ts
