+++ src/createContext.ts
@@ -3,13 +3,13 @@
 import { GLOBAL, GLOBAL_KEY_TO_ACCESS_NATIVE_STORAGE } from './constants'
 import { AsyncScheduler } from './schedulers'
 import * as non_thunk_list from './stdlib/list'
-import * as thunk_list from './stdlib/thunk_list'
+import * as thunk_list from './stdlib/thunk-list'
 import { list_to_vector } from './stdlib/list'
 import { listPrelude } from './stdlib/list.prelude'
 import * as misc from './stdlib/misc'
 import * as parser from './stdlib/parser'
 import * as non_thunk_stream from './stdlib/stream'
-import * as thunk_stream from './stdlib/thunk_stream'
+import * as thunk_stream from './stdlib/thunk-stream'
 import { streamPrelude } from './stdlib/stream.prelude'
 import { Context, CustomBuiltIns, Value, Variant } from './types'
 import * as operators from './utils/operators'
