+++ src/interpreter/environment-utils.ts
