+++ src/interpreter/evaluators.ts
@@ -5,10 +5,10 @@ import { Context, Value } from '../types'
 import { evaluateBinaryExpression, evaluateUnaryExpression } from '../utils/operators'
 import * as rttc from '../utils/rttc'
 import Closure, { BoundedApplyFunction } from './closure'
-import * as env from './environmentUtils'
-import { BreakValue, ContinueValue, ReturnValue, TailCallReturnValue } from './evaluatorUtils'
-import { Evaluator, getArgs, apply } from './evaluatorUtils'
-import { transformLogicalExpression, reduceIf, evaluateBlockStatement } from './evaluatorUtils'
+import * as env from './environment-utils'
+import { BreakValue, ContinueValue, ReturnValue, TailCallReturnValue } from './evaluator-utils'
+import { Evaluator, getArgs, apply } from './evaluator-utils'
+import { transformLogicalExpression, reduceIf, evaluateBlockStatement } from './evaluator-utils'
 import { makeThunkAware } from './thunk'
 
 export function getEvaluators(
