+++ src/stdlib/thunk-list.ts
